(improv) rename app folder to redux
