+++ webview/features/requestAuth/BasicAuth/index.tsx
@@ -1,6 +1,6 @@
 import * as React from "react";
 import "./styles.css";
-import { useAppDispatch, useAppSelector } from "../../../app/hooks";
+import { useAppDispatch, useAppSelector } from "../../../redux/hooks";
 import {
   requestAuthOptionsUpdated,
   selectBasicAuthOptions,
