+++ webview/features/requestAuth/BearerToken/index.tsx
@@ -1,5 +1,5 @@
 import * as React from "react";
-import { useAppDispatch, useAppSelector } from "../../../app/hooks";
+import { useAppDispatch, useAppSelector } from "../../../redux/hooks";
 import {
   requestAuthOptionsUpdated,
   selectBearerAuthOptions,
